Defined initial Linux-oriented dotfiles
This is an attempt to do a _ground-up_, bash-centric dotfiles after
reading @webpro's article [1]. The notion of an _attempt_ is that
the hope is to simplify and start out by culling brew or Mac specific
aspects since my goal is to normalize Linux environments that I work
within.

I bought over most of the files from @webpro's [2] `system/`, along
with his `./install.sh` and '/bin`.

I did not fork Lars' repository - though I consider it my origin.

[1] https://medium.com/@webprolific/getting-started-with-dotfiles-43c3602fd789#.enb9islys
[2] https://github.com/webpro/dotfiles